+++ Source/GameLoop/Game.h
@@ -12,7 +12,6 @@ private:
 	MT::Renderer *renderer = nullptr;
 	SDL_Event event = {};
 	std::unique_ptr<UI> ui = nullptr;
-	Slider* sl = nullptr;
 
 public:
 	void Start();