+++ CHANGELOG.md
@@ -216,3 +216,4 @@ was removed or the new one was added to folder
 - New rendering function for advanced 2D shadows
 - Upr version for advanced shadows
 - Camera interace and math has been improved
+- Improved UiElemBase infrastructure